move task filter into scheduler
